Reference

SparseArraysBase.SparseArrayDOK — Type

SparseArrayDOK{T,N,F} <: AbstractSparseArray{T,N}

N-dimensional sparse Dictionary-of-keys (DOK) array with elements of type T, optionally with a function of type F to instantiate non-stored elements.

SparseArraysBase.SparseArrayDOK — Method

SparseArrayDOK{T}(undef, dims...[; getunstored])
SparseArrayDOK{T,N}(undef, dims...[; getunstored])

Construct an uninitialized N-dimensional SparseArrayDOK containing elements of type T. N can either be supplied explicitly, or be determined by the length or number of dims.

SparseArraysBase.sparse — Method

sparse(storage::Union{AbstractDict,AbstractDictionary}, dims...[; getunstored])

Construct an N-dimensional SparseArrayDOK containing elements of type T. Both T and N can either be supplied explicitly or be determined by the storage and the length or number of dims. If dims aren't specified, the size will be determined automatically from the input indices.

This constructor does not take ownership of the supplied storage, and will result in an independent container.

SparseArraysBase.sparserand — Function

sparserand([rng], [T::Type], dims; density::Real=0.5, randfun::Function=rand) -> A::SparseArrayDOK{T}

Create a random size dims sparse array in which the probability of any element being stored is independently given by density. The optional rng argument specifies a random number generator, see also Random. The optional T argument specifies the element type, which defaults to Float64. The optional randfun argument can be used to control the type of random elements, and should support the signature randfun(rng, T, N) to generate N entries of type T.

See also sparserand!.

SparseArraysBase.sparserand! — Function

sparserand!([rng], A::AbstractArray; density::Real=0.5, randfun::Function=rand) -> A

Overwrite part of an array with random entries, where the probability of overwriting is independently given by density. The optional rng argument specifies a random number generator, see also Random. The optional randfun argument can be used to control the type of random elements, and should support the signature randfun(rng, T, N) to generate N entries of type T.

See also sparserand.

SparseArraysBase.sparsezeros — Function

sparsezeros([T::Type], dims[; getunstored]) -> A::SparseArrayDOK{T}

Create an empty size dims sparse array. The optional T argument specifies the element type, which defaults to Float64.