Reference

ITensorFormatter.main — Method

NAME
       ITensorFormatter.main - format Julia source code
       ITensorPkgFormatter.main - format ITensor Julia packages and generate READMEs

SYNOPSIS
       julia -m ITensorFormatter [<options>] <path>...
       julia -m ITensorPkgFormatter [<options>] <path>...

DESCRIPTION
`ITensorFormatter.main` (typically invoked as `julia -m ITensorFormatter` or `itfmt`)
formats Julia source code, Project.toml, and optionally YAML files using the ITensorFormatter.jl formatter.

`ITensorPkgFormatter.main` (typically invoked as `julia -m ITensorPkgFormatter` or `itpkgfmt`)
performs all formatting as above, and additionally generates README documentation for each provided ITensor package directory.

OPTIONS
<path>...
Input path(s) (files and/or directories) to process. For directories,
all files (recursively) with the '*.jl' suffix are used as input files.
--help
Print this message.
--version
Print ITensorFormatter and julia version information.
--yaml
Also format YAML files (*.yml, *.yaml). Disabled by default.
        

ITensorFormatter.make_index! — Method

make_index!(pkgroot::AbstractString; inputfile, outputdir, flavor, name, postprocess)

Generate docs/src/index.md from examples/README.jl using Literate.markdown.

ITensorFormatter.make_readme! — Method

make_readme!(pkgroot::AbstractString; inputfile, outputdir, flavor, name, postprocess)

Generate README.md from examples/README.jl using Literate.markdown.

ITensorFormatter.ITensorPkgFormatter.main — Method

NAME
       ITensorFormatter.main - format Julia source code
       ITensorPkgFormatter.main - format ITensor Julia packages and generate READMEs

SYNOPSIS
       julia -m ITensorFormatter [<options>] <path>...
       julia -m ITensorPkgFormatter [<options>] <path>...

DESCRIPTION
`ITensorFormatter.main` (typically invoked as `julia -m ITensorFormatter` or `itfmt`)
formats Julia source code, Project.toml, and optionally YAML files using the ITensorFormatter.jl formatter.

`ITensorPkgFormatter.main` (typically invoked as `julia -m ITensorPkgFormatter` or `itpkgfmt`)
performs all formatting as above, and additionally generates README documentation for each provided ITensor package directory.

OPTIONS
<path>...
Input path(s) (files and/or directories) to process. For directories,
all files (recursively) with the '*.jl' suffix are used as input files.
--help
Print this message.
--version
Print ITensorFormatter and julia version information.
--yaml
Also format YAML files (*.yml, *.yaml). Disabled by default.