Reference

KroneckerArrays.AbstractKroneckerArray — Type

abstract type AbstractKroneckerArray{T, N} <: AbstractArray{T, N} end

Abstract supertype for arrays that have a kronecker product structure, i.e. that can be written as AB = A ⊗ B.

KroneckerArrays.CartesianPair — Type

CartesianPair(a, b)

Represents a single element, the cartesian product of two arbitrary objects a and b.

KroneckerArrays.CartesianProduct — Type

CartesianProduct(a::AbstractVector, b::AbstractVector)

Represents the cartesian product of two collections a and b.

KroneckerArrays.CartesianProductUnitRange — Type

CartesianProductUnitRange(a::AbstractUnitRange, b::AbstractUnitRange, range::AbstractUnitRange{T}) <: AbstractUnitRange{T}

Similar to CartesianProductVector, this represents the cartesian product of two ranges a and b. However, as a range it will behave as range, rather than CartesianPairs of the elements of a and b.

KroneckerArrays.CartesianProductVector — Type

CartesianProductVector(a::AbstractVector, b::AbstractVector, values::AbstractVector{T}) <: AbstractVector{T}

Similar to the CartesianProduct, this represents the cartesian product of two collections a and b. However, as a vector it will behave as values, rather than CartesianPairs of the elements of a and b.

KroneckerArrays.:× — Function

×(args...)
times(args...)

Construct an object that represents the Cartesian product of the provided args. By default this constructs the singular CartesianPair for unknown values, while attempting to promote to more structured types wherever possible. See also CartesianProduct, CartesianProductVector and CartesianProductUnitRange.

KroneckerArrays.:⊗ — Function

⊗(args...)
otimes(args...)

Construct an object that represents the Kronecker product of the provided args.

KroneckerArrays.cartesianproduct — Function

cartesianproduct(a::AbstractVector, b::AbstractVector, [values::AbstractVector])::AbstractVector

Construct an AbstractVector that represents the cartesian product a × b, but behaves as values. This behaves similar to ×, but forces promotion to a AbstractVector.

KroneckerArrays.cartesianrange — Function

cartesianrange(a::AbstractUnitRange, b::AbstractUnitRange, [range::AbstractUnitRange])::AbstractUnitRange

Construct a UnitRange that represents the cartesian product a × b, but behaves as range. This behaves similar to ×, but forces promotion to a AbstractUnitRange.

KroneckerArrays.kroneckerfactors — Function

kroneckerfactors(x) -> Tuple
kroneckerfactors(x, i) = kroneckerfactors(x)[i]

Extract the factors of x, where x is an object that represents a lazily composed product type.

KroneckerArrays.kroneckerfactortypes — Function

kroneckerfactortypes(x) -> Tuple
kroneckerfactortypes(x, i) = kroneckerfactortypes(x)[i]

Extract the types of the factors of x, where x is an object or type that represents a lazily composed product type.

KroneckerArrays.unproduct — Function

unproduct(a)

For an object that holds a cartesian product of indices and their corresponding values, this function removes the cartesian product layer and returns only the values.