Reference

BackendSelection.Algorithm — Type

BackendSelection.Algorithm

A type representing a backend for a function.

For example, a function might have multiple backends implementations, which internally are selected with a BackendSelection.Algorithm type.

This allows users to extend functionality with a new implementation but use the same interface.

BackendSelection.Backend — Type

BackendSelection.Backend

A type representing a backend for a function.

For example, a function might have multiple backends implementations, which internally are selected with a BackendSelection.Backend type.

This allows users to extend functionality with a new implementation but use the same interface.

BackendSelection.@Algorithm_str — Macro

@Algorithm_str

A convenience macro for writing Algorithm types, typically used when adding methods to a function that supports multiple backends.

BackendSelection.@Backend_str — Macro

@Backend_str

A convenience macro for writing Backend types, typically used when adding methods to a function that supports multiple backends.